Address: Split Address into multiple classes
Instead of just storing the full address in a single
string, the full address has been split into 4 different
sub-classes:
 * Block
 * Street
 * Unit
 * PostalCode
This change will help in sorting the address by any part of
the address.

As part of this change, the member {value} has been removed.
The toString() method now parses and returns the String that
{value} used to represent.

AdaptedPerson: instance where it references Address: {value}
is changed to reference the toString() method instead.