Add Logic for save failure
When a user fails to save his addressbook, ideally we want him to have the ability to try again and not simply exit. To do so, refactored isExit() to isSuccessfulExit() so run loop continues to run until SUCCESSFUL termination. The backend is handled by some simple and readable boolean logic, nothing fancy.